fix code mistake
var in this case is hoisted to the top of the function and will blow
out any passed value for ‘e’. In order to maintain code integrity,
don’t re-var ‘e’ since it has already been declared.  Now we could
theoretically pass a value for ‘e’ and have it work correctly.